+++ apps/AppleApp/Brownfield Apple App/components/ContentView.swift
@@ -23,10 +23,23 @@ struct ContentView: View {
 
                 MessagesView()
 
-                ReactNativeView(moduleName: "main")
+                ReactNativeView(
+                    moduleName: "main",
+                    initialProperties: [
+                        "nativeOsVersionLabel":
+                            "\(UIDevice.current.systemName) \(UIDevice.current.systemVersion)"
+                    ]
+                )
                     .navigationBarHidden(true)
                     .clipShape(RoundedRectangle(cornerRadius: 16))
                     .background(Color(UIColor.systemBackground))
+                
+                Button("Stop React Native") {
+                    ReactNativeBrownfield.shared.stopReactNative()
+                }
+                    .buttonStyle(PlainButtonStyle())
+                    .padding(.top)
+                    .foregroundColor(.red)
             }
             .frame(maxWidth: .infinity, maxHeight: .infinity)
             .padding(16)